+++ battlecode22-scaffold/autoscript.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 start_time=$SECONDS
-team1=AMapBot
-team2=OSplitBot
+team1=APreQualBot
+team2=OMapBot
 # team2=OPerfectProducerBot
 
 mkdir -p logs